agents/gita-guide/requirements.txt
anthropic>=0.18.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
